rename react components files
