+++ src/screens/newtab/components/Background.tsx
