+++ src/screens/newtab/components/Clock.tsx
